Overlay gradient on second pass. Partial failure.
Intended to set number of event triggers as an
attribute. Failed to implement as intended.
Commiting here so I can see what I did for future
reference.

Other components of the overlay are working as
intended so far. the attribute method for counting
will be removed, but the rest of the changes are
part of the needed script for the next feature.